requirements.txt
# Core FastAPI and web framework
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0

# OpenAI and AI libraries
openai>=1.0.0
tiktoken>=0.4.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Data processing and ML
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0

# HTTP and networking
requests>=2.28.0
httpx>=0.24.0

# Logging and utilities
loguru>=0.7.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# AWS (optional)
boto3>=1.26.0

# Vector database (optional)
# weaviate-client>=3.15.0

# MCP (optional)
# mcp>=1.0.0
# websockets>=12.0
# asyncio-mqtt>=0.16.0
